Design, Chore: 상세 내용 모달 컴포넌트 화 #1, #20
