+++ src/components/ModalAddFriend.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const ModalAddFriend = () => {
+
+    
+  return (
+    <div>ModalAddFriend</div>
+  )
+}
+
+export default ModalAddFriend
